[snmp] add hexstring channel datatype (#6070)
