+++ bundles/org.openhab.binding.snmp/src/main/java/org/openhab/binding/snmp/internal/SnmpDatatype.java
@@ -24,5 +24,6 @@ public enum SnmpDatatype {
     COUNTER64,
     FLOAT,
     STRING,
+    HEXSTRING,
     IPADDRESS
 }
